requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
deprecation==2.1.0
Faker==24.0.0
gotrue==2.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httpx==0.25.2
hyperframe==6.0.1
idna==3.7
numpy==2.0.1
packaging==24.1
pandas==2.2.3
ply==3.11
postgrest==0.15.1
protobuf==5.26.1
pydantic==2.8.2
pydantic_core==2.20.1
pyenchant==3.2.2
PyJWT==2.8.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt6==6.7.0
PyQt6-Charts==6.7.0
PyQt6-Charts-Qt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pyqtspinner==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
qasync==0.27.1
realtime==1.0.6
requests==2.31.0
sip==6.7.11
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websockets==12.0
bcrypt==4.2.0
python_dotenv==1.1.0
